# PERF: Faster transform direction if already passed an Enum