Currently, `Enum.create()` adds quite a bit of overhead to the transform calls, so we probably don't want to call it when unnecessary. Here I added `isinstance()` checks to handle it in a completely backwards compatible way.

```python
In [1]: from pyproj import Transformer

In [2]: transformer = Transformer.from_crs(2263, 4326, always_xy=True)

In [3]: %timeit transformer.transform(1, 2, direction="forward")
7.96 µs ± 32 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [4]: %timeit transformer.transform(1, 2, direction="FORWARD")
4.18 µs ± 33.9 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [5]: %timeit transformer.transform(1, 2)
3.26 µs ± 21.6 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

 I think it might be even nicer to remove `Enum.create()` entirely and move over to a `StrEnum` / `BaseEnum(str, Enum)` and then the dictionary lookup could be either `"FORWARD"` or `TransformDirection.FORWARD` with the caveat that you'd lose the case insensitive conversion, but the code would be a bit cleaner everywhere then.