@snowman2 2022-12-18
StrEnum looks nice, but is Python 3.11+.