@snowman2 2022-12-18
I wonder if it would be simpler to add the change there: https://github.com/pyproj4/pyproj/blob/e345729e4283535acf43da40e3679a4e8fae6152/pyproj/enums.py#L14

```python
if isinstance(item, cls):
    return item
```