@greglucas 2022-12-18
Adding the change directly to Enums would be cleaner, but does add a bit of a speed hit still... (I assume going from Cython -> Python -> Cython adds some overhead?)


> StrEnum looks nice, but is Python 3.11+.

I think we could get away with just subclassing string here and defining:
`class TransformDirection(str, BaseEnum)`
there isn't much more to the stdlib implementation:
https://github.com/python/cpython/blob/1cf3d78c92eb07dc09d15cc2e773b0b1b9436825/Lib/enum.py#L1263

I think the question if you want to go this route is whether losing case sensitivity is a deal breaker or not?