@snowman2 2022-12-19
To make the code simpler, thoughts about adding this function:
```python
cdef PJ_DIRECTION get_pj_direction(object direction) except *:
    if not isinstance(direction, TransformDirection):
        direction = TransformDirection.create(direction)
    return _PJ_DIRECTION_MAP[direction]
```
and doing this:
```python
cdef PJ_DIRECTION pj_direction = get_pj_direction(direction)
```