@snowman2 2022-12-19
> I think we could get away with just subclassing string

Have you tested the performance of this implementation?