@greglucas 2022-12-19
> To make the code simpler, thoughts about adding this function:

:+1: That seems like it would be nice!

However, I think we should decide on the next comment first.

> >  I think we could get away with just subclassing string

> Have you tested the performance of this implementation?

Yes, it is actually faster because you don't need an isinstance check. You can go right into your dictionary with both `TransformDirection.FORWARD` and the string `"FORWARD"` because `"FORWARD" == TransformDirection.FORWARD` and both versions will match the key. (thus negating the need for the lookup above)

Downsides are that `"forward" != TransformDirection.Forward` so random cased strings won't work (probably good for Enums actually) and the nice error message of valid options are no longer there.