@snowman2 2022-12-19
How much of a performance improvement do you get compared to this PR? I think that if it is significant, I don't think that being case insensitive is as important.