@snowman2 2022-12-19
I am wondering if this would be faster and still allow the error message and support case insensitive input:

```python
cdef PJ_DIRECTION get_pj_direction(object direction) except *:
    try:
         return _PJ_DIRECTION_MAP[direction]
    except KeyError:
        direction = TransformDirection.create(direction)
        return _PJ_DIRECTION_MAP[direction]
```