@greglucas 2022-12-19
I like your most recent suggestion of the try/except inside a helper function. We can add `StrEnum` later too if desired.